Update the repo of recipes on `quelpa-upgrade` or `quelpa-self-upgrade`.
The documentation for `quelpa-update-melpa-p` indicates that when nil certain
interactive commands will update the MELPA repo, but this is not the case. This
commit fixes that problem by making the checkout function take a parameter
instead of reading a dynamic variable. This means the setup function maintains
its checkout behavior with regards to `quelpa-update-melpa-p` but the two
commands mentioned above will always do an update. Notably, this update occurs
only once, even though `quelpa-upgrade` calls `quelpa-self-upgrade`.