Allow create-tar to signal error on failure
I've added quelpa-build--create-tar to signal an error in the event of receiving
return code 2, which, according to the tar(1) manual, indicates a fatal error.
Apart from that, I've also allowed it to signal 1 and continue, as well as the
usual 0.

I've seen someone complaining online that they received a return value of 5,
which is weird beyond belief, but I've added a default value as well, along with
the ability to view the return value, in case that's useful.