Made create-tar use relative directories rather than absolute ones
As stated in the Elisp Manual
here (https://www.gnu.org/software/emacs/manual/html_node/elisp/Synchronous-Processes.html),
under `process-files':

> To avoid confusion, it may be best to avoid absolute file names in args, but
> rather to specify all file names as relative to default-directory. The
> function file-relative-name is useful for constructing such relative file
> names.

Now, what this function does is gets the `file-truename' of both file and dir,
and then uses `file-relative-name', with `dest-dir' as the root, before setting
`default-directory' to `dest-dir'.

There will definitely be a concern however that this function will break in the
event we use multiple `files', so that's next on the list.