Revamp code for managing create-tar with dir or files
In order to forestall the problem I foresaw in commit 94955512, I've made sure
that in the event that dir is `nil' and files is a list (e.g. when we have a
MELPA recipe that specifies :files), what will happen is that quelpa will copy
these files into a temporary directory, pack everything into a TAR file, and
then use that temporary directory instead.

That's what `quelpa-build--stage-files' is for: it takes `dir' and an optional
`files' list (much like `quelpa-build--create-tar'), and then returns either
`dir' (if `dir' isn't nil), or the path of a temporary directory that is used to
stage the many different files (referenced by the file's `file-truename').

Based on the documentation on MELPA recipes[1], I don't expect to see /any/
problems, but I have a suspicion if 1) the package files require a specific
directory structure, and use a :files directive, this might break. But trying to
forestall that problem looks like it's too much work, and I hope I don't need to
do it, judging by how many packages I've seen appear to be, at most complex, a
list of files in one directory, hopefully this won't be a problem.

Hopefully.

[1]: https://github.com/melpa/melpa#recipe-format