Fix broken quelpa-build--stage-files & create-tar
So it turns out that, apart from a howler of an error in the let-form of
`quelpa-build--create-tar', I apparently didn't understand the nature of how
`'tar' works (protip: --directory is applied sequentially, so you need to apply
it BEFORE -cvf).

Apart from that, I also didn't understand how package .tar files worked, either
-- quelpa-build--create-tar was popping out unusable .tar files. Turns out that
part of it was fixed by moving --directory to the beginning of the options, but
what also needed to happen was that .tar files needed to have the name of the
folder, which is usually the name of the package plus the version number.

That meant also making sure that quelpa-build--stage-files also needed to be
modified, to include a new argument, PKG-LABEL. And then, rather than doing that
awkward string comparison, we might as well just check to see if FILES was nil.
If it was, then we create a directory in staging-dir (named PKG-DIR), change
staging-dir to that directory, and copy all the files listed into staging-dir,
before proceeding as normal.

So far it works okay with installing a package using a git-based fetcher. More
extensive testing needed.